pyproject.toml
[project]
name = "web-dev"
version = "0.1.0"
description = "web dev projects"
authors = [
    {name = "swathisivaprabu",email = "swathisivaprabu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "datetime (>=5.5,<6.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
